[Ingest Manager] Add .catch to handle Promise rejection (#80056)
## Summary

Prevent an unhandled rejection (which crashes Kibana) when installing a package. Initially found as a part of https://github.com/elastic/kibana/issues/79560#issuecomment-704581602 & [this PR](https://github.com/elastic/kibana/pull/79791/files#diff-e2fac97e334155d887ecee2a8434120f9150f65c85725c4a25f261da79233cf0R336-R340) to fix it. 

While the conditions of https://github.com/elastic/kibana/issues/79560 are a bit of an edge case, we could encounter this behavior on any expected error in this promise chain.

 * Add .catch to prevent unhandled promise rejection warning/crash.
 * pulled `installPackage` out to a separate file so it's easier to test
 * added a test for this specific error where an error in `installIndexPatterns` or `installKibanaAssets` would ultimately crash Kibana. More detail in the comments https://github.com/elastic/kibana/blob/608fc3ef1b1306fc0dd76ff07b6652c221ed242b/x-pack/plugins/ingest_manager/server/services/epm/packages/_install_package.ts#L81-L95



### Checklist
- [x] [Unit or functional tests](https://www.elastic.co/guide/en/kibana/master/development-tests.html) were updated or added to match the most common scenarios
